+++ CHANGELOG.rst
@@ -7,8 +7,8 @@ v0.3.0
 | **Release-date:** 15 Jun, 2023
 | **Release-by:** Asif Saif Uddin
 
-- Drop Python 2 support, remove six.
-- Uses `PromptSession()` class from prompt_toolkit instead of prompt() function (#63).
+- Drop Python 2 support, remove :py:mod:`~six`.
+- Uses ``PromptSession()`` class from prompt_toolkit instead of ``prompt()`` function (#63).
 - Added filter for hidden commands and options (#86).
 - Added click's autocompletion support (#88).
 - Added tab-completion for Path and BOOL type arguments (#95).